move console service to jacdac
