+++ libs/jacdac-drivers/pxt.json
@@ -5,7 +5,6 @@
         "sensorclient.ts",
         "buttonclient.ts",
         "lightsensorclient.ts",
-        "console.ts",
         "battery.ts",
         "accelerometerclient.ts",
         "touchclient.ts",
