+++ libs/jacdac/console.ts
@@ -1,5 +1,6 @@
 namespace jacdac {
-    class ConsoleClient extends Client {
+    //% fixedInstances
+    export class ConsoleClient extends Client {
         public suppressForwading: boolean;
         constructor() {
             super("log", jacdac.LOGGER_DEVICE_CLASS);
@@ -30,16 +31,39 @@ namespace jacdac {
         }
     }
 
-    let _consoleClient: ConsoleClient;
+    //% fixedInstance whenUsed block="console"
+    export const consoleClient = new ConsoleClient();
+
     /**
-     * Sends console messages over JacDac
+     * Receives messages from console.log
      */
-    //% blockId=jacdac_broadcast_console block="jacdac broadcast console"
-    //% group="Console"
-    export function broadcastConsole() {
-        if (!_consoleClient) {
-            _consoleClient = new ConsoleClient();
-            _consoleClient.start();
+    //% fixedInstances
+    export class ConsoleService extends Service {
+        constructor() {
+            super("log", jacdac.LOGGER_DEVICE_CLASS); // TODO pickup type from DAL
+        }
+
+        public handlePacket(pkt: Buffer): boolean {
+            const packet = new JDPacket(pkt);
+            const packetSize = packet.size;
+            if (!packetSize) return true;
+
+            const priority = packet.data.getNumber(NumberFormat.UInt8LE, 0);
+            // shortcut
+            if (priority < console.minPriority) return true;
+
+            // send message to console
+            let str = "";
+            for (let i = 1; i < packetSize; i++)
+                str += String.fromCharCode(packet.data.getNumber(NumberFormat.UInt8LE, i));
+
+            // pipe to console TODO suppress forwarding
+            console.add(priority, str);
+
+            return true;
         }
     }
+
+    //% fixedInstance whenUsed block="console service"
+    export const consoleService = new ConsoleService();
 }
